Add upload a photo
This enables uploading a photo to firestore which is the first step in
processing it and all of its companions.

I was using a naming convention of all images but it seems the tutorial
has some weird differentiation between an image and a photo so I've
moved it all inline with the naming used in the tutorial to keep it
logical.