Show our images in home
Gives us a view to our own content which is nice. One catch with it is
they're displayed in chronological order of upload. Look that's neat but
not very intuitive for the user. It would be nice to have a `createdAt`
field on the photos but that seemed beyond the scope of this change.
Instead I've just added an in place javascript reverse. I expect this
reverse operation will have a time complexity of O(n) which does offer
an opportunity for optimisation as we grow in size of user content we
expect.