Remove eslint from firebase functions
So the tutorial didn't add eslint and their code had some errors. That
posed a question for do I fix the code or go to be closer to the
tutorial and considering this is really just about me following the
tutorial I went for that. In future I'll be using eslint but this
project is really about getting in the deep end and trying something
out.