Add display of doppelganger pictures
I removed the JS reverse on sort order, currently the order is really
based of the random ids and they aren't linear so it's kind of funny to
reverse an meaningless order.

I considered fixing the warning for all of the unnecessary `\`
characters escaping on `src/features/Home.js:16` but that line is long
and I don't think is really valuable for me to dig into it at this
point.