Add project specific getting started instructions.
This is to give people a start into using and navigating the project.

I considered testing out the commands but I don't yet expect anyone to
use this it's more for my succinctness. I would expect others in the
future to extend on this as they get into the project.

There will always be the original tutorial which I've linked to to fall
back on.